refactor: cut back to 2 modules
