--- src/uvtrick/__init__.py
@@ -1,5 +1,0 @@
-from .core import load
-from .env import Env
-from .trick import uvtrick_
-
-__all__ = ("load", "Env", "uvtrick_")
